Collect all test cases automatically.  (#17717)
What is the motivation for this PR?
We aim to gather a comprehensive list of all test cases and determine the topologies they are compatible with. In this PR, we use pytest to collect all test cases while simultaneously analyzing pytest marks in each test script to extract topology information. Additionally, we set up a pipeline to perform this collection daily and store the results in Kusto.

How did you do it?
In this PR, we use pytest to collect all test cases while simultaneously analyzing pytest marks in each test script to extract topology information. Additionally, we set up a pipeline to perform this collection daily and store the results in Kusto.

How did you verify/test it?
You can refer to https://dev.azure.com/mssonic/internal/_build/results?buildId=838867&view=results, after merging this PR, the collection of sonic-mgmt will success.