Merged PR 12259: Fix PTF to VM connectivity issue for TK5 lab
Fix PTF to VM connectivity issue for TK5 lab

In our lab, mgmt interface of PTF is configured with lab mgmt network IP. This is because many tests requrie DUT to reach PTF via mgmt network. For example the NTP and TACACS testing. PTF is configured as NTP and TACACS server.

Neighbor VMs are configured with 172 subnet private IP for mgmt interface. It's because we have large number of neighbor VMs. Simply there is not enough lab mgmt network IP for each of them.

For PTF container to reach neighbor VMs, the packets orignated from PTF must reach an interface awares of the 172 subnet. Usually we configure the br1 bridge on test server with a lab mgmt network IP and a 172 subnet IP as secondary IP. That makes it possible for the PTF to reach neighbor VM's 172 subnet IP.

In TK5 lab, the default GW in PTF is set to the M0 vlan interface IP (10.212.65.1). Unfortunately the TK5 M0 vlan interface does not have the 172 subnet IP configured. If ping 172 subnet IP from PTF, the packets would be forwarded to the M0 vlan interface (10.212.65.1) and then dropped.

To make this work, we must set the default GW in PTF to the mgmt IP of lab server where the PTF container is running. Then if ping 172 subnet IP from PTF container, the packets will be sent to the mgmt IP of lab server. Then the lab server can forward it to 172 subnet because br1 is configured with 172 subnet IP too.

Thanks to Ryan Garthwaite for reporting and help debugging this issue.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix addressing VM connectivity issues in the TK5 lab.

#### PR Summary
The pull request corrects the management gateway IP addresses in the ansible host variable files for TK5 lab servers to resolve PTF to VM connectivity issues.
- `ansible/host_vars/STRTK5-SERV-03.yml`: Updated mgmt_gw from 10.212.65.1 to 10.212.65.4.
- `ansible/host_vars/STRTK5-SERV-04.yml`: Updated mgmt_gw from 10.212.65.1 to 10.212.65.5.
- `ansible/host_vars/STRTK5-SERV-05.yml` & `STRTK5-SERV-06.yml`: Changed mgmt_gw to 10.212.65.6 and 10.212.65.7, respectively.
- `ansible/host_vars/STRTK5-SERV-07.yml` to `STRTK5-SERV-10.yml`: Modified mgmt_gw values from 10.212.65.1 to 10.212.65.8, 10.212.65.9, 10.212.65.10, and 10.212.65.11 respectively.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->