[TESTGAP] Implement new test case to check the crm nexthop_group (#18465)
This is a new test case to verify the CRM nexthop_group threshold, which is typically set to at least 256. However, for some HWSKUs, such as 7215 and 720dt, the value is lower. This is acceptable due to their specific application scenarios, which have lower requirements for ECMP group capacity.

What is the motivation for this PR?
To address the test gap related to nexthop group requirements.

How did you do it?
To verify the total number of nexthop groups supported by the platform and ensure it meets the minimum required threshold.

How did you verify/test it?
Run the test case on device.

Any platform specific information?
Supported testbed topology if it's a new test case?
't0', 't1', 'm0', 'mx', 'm1', 'm2', 'm3'

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>