Recovery for inconsistent mux state issue (#18342)
What is the motivation for this PR?
To improve the speed of mux state recovery by restarting the linkmgrd process directly, reducing the time taken to recover from an inconsistent mux state.

How did you do it?
Added a linkmgrd restart command to save time for inconsistent mux state.

How did you verify/test it?
Ran the pre-test and confirmed that it can recover the inconsistent mux state.

Signed-off-by: yenlu <yenlu@cisco.com>