[CRM_AVAILABLE] To add delay when read the CRM info (#18517)
In some cases, the CRM info is not available immediately after the test startup, need to wait for some time before read it, otherwise, the test case could run into failure with.

      pytest.fail(f"Resource '{resource_name}' not found in CRM resources output on platform '{hwsku}'.")
E Failed: Resource 'nexthop_group' not found in CRM resources output on platform 'force10-s6000'.

What is the motivation for this PR?
To fix the flaky of the case due to timing issue.

How did you do it?
To add wait_until when read the CRM info from DUT.

How did you verify/test it?
Run test on platform.