fix: update BFD check result after initial failure (#18498)
Description of PR
This is a follow up PR for #18466 to handle the BFD status update in a more proper way.

Summary:
Fixes # (issue) Microsoft ADO 32945039

Approach
What is the motivation for this PR?
Since we are checking the BFD count of all ASICs in a multithreading way, we don't want to always update check_result["failed"] at the beginning of _check_bfd_up_count() as it might lead to a race condition scenario.

How did you do it?
How did you verify/test it?
I did the following verifications:

BFD status of all ASICs is as expected when the sanity check starts: https://elastictest.org/scheduler/testplan/682c37be20f2315851bb82ae
Manually verified that if an ASIC’s BFD status is not expected at the start of the sanity check but changes to the expected count during the wait_until period, the sanity check will pass
Manually verified that if an ASIC’s BFD status is always not expected, the sanity check will fail
Any platform specific information?

co-authorized by: jianquanye@microsoft.com