Adding fix for m20 test (#18531)
Description of PR
Currently, the results were verified using round which resulted in some failures where 10.85% rounds to 11 and fails the test on Nokia Platforms. Modified the condition to have some more room for values of drop between 9-11%.

Summary:
Fixes # (issue)
#18043

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202012
 202205
 202305
 202311
 202405
 202411
Approach
What is the motivation for this PR?
To enhance the check for verification of drops in the test case.

How did you do it?
By changing the assertion used to verify the drops.

How did you verify/test it?
Tested on Nokia Chassis in msft lab

Co-authored-by: yatishkoul <yatishkoul@gmail.com>
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>