Add a parameter to get-impacted-area (#18589)
Description of PR
Previously, we hardcoded origin/master to determine the differences between the base branch and the current branch. This approach was sufficient when the impacted area logic was only used on the master branch. However, as we now plan to support impacted area detection on additional branches, we need to replace this hardcoded reference with the actual target build branch.

Summary:
Fixes # (issue)

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202205
 202305
 202311
 202405
 202411
 202505
Approach
What is the motivation for this PR?
Previously, we hardcoded origin/master to determine the differences between the base branch and the current branch. This approach was sufficient when the impacted area logic was only used on the master branch. However, as we now plan to support impacted area detection on additional branches, we need to replace this hardcoded reference with the actual target build branch.

How did you do it?
Change this hardcode to the actual target build branch.

How did you verify/test it?
As the target build branch of this PR is mater branch, after we change the hardcode to the actual target build branch, it can successfully get the name of the target build branch.
image

co-authorized by: jianquanye@microsoft.com

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>