Updating telemetry metrics interfaces for metrics collection (#17973)
What is the motivation for this PR?
The current telemetry metrics interfaces is lacking the common metrics label and names for data collection, such as port/queue/psu/sensor and etc.

How did you do it?
This PR adds the most frequently used metrics labels as well as updating the examples for usage.

How did you verify/test it?
Validated with local influx db deployment.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>