add dualtor test test_mux_forwarding_state_consistency (#18232)
What is the motivation for this PR?
To verify that the forwarding state is corrected when socnet agent randomly resets the admin state on host side.

How did you do it?
By setting the ToR's desired mux state (standby or active) and using nic_simulator to inject an incorrect state. The test then checks the state restores.

How did you verify/test it?
The forwarding state consistently matches with the desired state in both standby and active scenarios.

Signed-off-by: yenlu <yenlu@cisco.com>