Modify gnmi stub generate script inline with protoc issue fixed (#18606)
Modify the generate gNMI stub script to check if the directory exists. This was due to a bug grpc/grpc#39583 in grpc_tools.protoc which is now fixed.

What is the motivation for this PR?
Fix issue with stub generation script.

How did you do it?
See description above.

How did you verify/test it?
Manually

Signed-off-by: yenlu <yenlu@cisco.com>