[ceos]: Optimize CEOS image by disabling RPM service (#18685)
What is the motivation for this PR?
According to the profiling data, we found the disk write operations will become a bottleneck if we start a large number of cEOS container as neighbors at the add-topo step. All most write operations come from the RPM service that will update the package list and upgrade packages during the starting stage.

This fix can reduce about 40% time (225mins to 137 mins) consumption for 32 neighbors creating.

How did you do it?
Comment the RPM service of start script of cEOS to disable the RPM service. Because to update package list and upgrade packages isn't necessary in most of our scenarios.

How did you verify/test it?
Check it locally and profiling data, there is no obvious write operations can be found.

Signed-off-by: Ze Gan <ganze718@gmail.com>
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>