[CRM AVAILABLE] To enhance the crm tests for TD3 and Cisco devices (#18733)
For TD3, the maximum number of nexthop groups is 255, which is below the minimum requirement of 256, resulting in test failures. To address this, TD3 should be added to SKU_NEXTHOP_THRESHOLDS.

Additionally, some HWSKUs require more time for the crm command to become operational. To accommodate this, the wait time has been increased from 90 seconds to 360 seconds in worst case.

What is the motivation for this PR?
To fix the flaky failure on Cisco devices and the consistent failure on TD3

How did you do it?
For TD3, add 255 as a threshold.
Extend maximum wait time from 90 seconds to 360 seconds.

How did you verify/test it?
Run test on TD3 and Cisco devices.

Signed-off-by: yenlu <yenlu@cisco.com>