[ceos_ensure_reachable.yml]: Fix wrong host of ceos (#18734)
What is the motivation for this PR?
The ansible-playbook checks container's reachability on the wrong host. If the number of neighbors is large, it can waste a lot of time.

How did you do it?
Fix the wrong host from ansible executing host to container host.

How did you verify/test it?
Check it locally.

Signed-off-by: Ze Gan <ganze718@gmail.com>
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>