Determine if the eos switch is running in multiagent mode, and stop/start the (#18748)
What is the motivation for this PR?
Addressing this bug means that the test bgp/test_bgp_gr_helper.py will restart the correct "bgpd" agent and pass irrespective of the version of cEOSLab being run on the docker peers.

How did you do it?
The test runs the command "show ip route summary | json" on the cEOSLab peer instance and parses the output to identify whether or not the instance is running in "multi-agent" mode. This information is then used to start/stop either the "Bgp" or "Rib" processes.

How did you verify/test it?
Verified by running the same bgp/test_bgp_gr_helper.py with and without the fix. With the fix the correct "bgpd" agent is restarted by the test.

Signed-off-by: Yael Tzur <ytzur@nvidia.com>