Changing Loss Percentage in test_m2o_oversubscribe_lossy.py (#18265)
Description of PR
In the test case test_m2o_oversubscribe_lossy.py, we generate two background traffic streams and two test flow traffic streams. The background traffic consists of two lossless streams, each utilizing 25% of the bandwidth. The test flow traffic includes lossy streams, each configured to consume either 30% bandwidth, or one at 25% and the other at 30%.

This results in an aggregate traffic load of 105% or 110%, intentionally oversubscribing the link. Under this setup, we expect approximately 10% packet loss due to congestion. The test currently verifies this by asserting that abs(drop_percentage - 10) < 1, effectively accepting a drop range of 9% to 11%. In most cases, the drop is not expected to exceed 10%.

However, during testing on Arista platforms, the observed drop percentages were slightly lower—approximately 8.74% and 8.76%. This behavior was discussed with @vmittal-msft and @amitpawar12. We recommend bringing this up in the upcoming Ixia Community meeting to align on expectations before proceeding with a merge.

Summary:
Fixes # (issue)

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement

Approach
What is the motivation for this PR?
To change the drop percentage for snappi test.

How did you do it?
Changed the drop percentage from 9-11 to 8-10%.

How did you verify/test it?
Tested in msft lab

Co-authored-by: yatishkoul <yatishkoul@gmail.com>
Signed-off-by: yenlu <yenlu@cisco.com>