[COPP] Check ptf reachability before running the copp tests (#18470)
This is to fix an issue that is caused by dut unreachability from ptf after a config reload of the dut in the copp test. The reason of the unreachability is we have multiple path from dut to ptf in our lab network. And after the reload, the path could change but the arp entries may not be all updated in all the network nodes in the new path.
If copp test try to establish socket for ptf_nn_agent between dut and ptf, and when the dut is not reachable, the ptf test command will hang there forever.
Add a ping before the ptf command, in one hand it will update the arp in the path to make sure the dut is reachable; in the other hand, it will fail the test immediately when the dut is somehow indeed unreachable and save the time for the timeout.