[test_bgp_stress_link_flap] update completeness_level for bgp link flap case (#18336)
What is the motivation for this PR?
We plan to use this case test_bgp_stress_link_flap for the longevity test, current thorough run time is not meet the requirement, need enlarge the run time value for thorough level.

How did you do it?
update completeness_level for test_bgp_stress_link_flap.py
Enlarge the runtime for the case and add the memory monitor during test in thorough level.

How did you verify/test it?
Run the pipeline

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>