[bgp] Fix test_ipv6_bgp_scale fails in multi-server (#18877)
What is the motivation for this PR?
test_ipv6_bgp_scale failed in multi servers because the test case trys to announce route to the VMs not belonging to it.

How did you do it?
In setup stage, construct the map of ptf_ip and corresponding dut interfaces connect to the ptf ports.
Then announce / withdraw routes based on above map.

How did you verify/test it?
Run tests

Signed-off-by: yenlu <yenlu@cisco.com>