Revise vrf definition to vrf instance
In EOS v4.21 the "vrf definition" command was deprecated, replaced by "vrf
instance".  In a forthcoming EOS version the "vrf definition" will be
removed from EOS altogether.  We need to change all of the ansible templates
to change the command to the supported "vrf instance" form.  The oldest
cEOS-lab available to download is 4.23 thus all available versions
support the new "vrf instance" form.