xfail test_iface_naming_mode.py::test_config_interface_speed (#18693)
Summary:
test_iface_naming_mode.py::test_config_interface_speed is failing due to a bug in the interpretation of config_db.json when determining the supported speeds of an intf.

xfailing until fix for issue is made to unblock other test cases in the test file. Low priority issue as only fixed config is used.

Signed-off-by: yenlu <yenlu@cisco.com>