Fix test_dynamic_acl.py by updating list of topos without LAGs (#18806)
What is the motivation for this PR?
The index of out bounds failure was a heavy hitter for us on a few topos.

How did you do it?
Added all known topos without LAGs to topos_no_portchannels tuple. This changes the code path and fetches the list of upstream ports as expected.

How did you verify/test it?
Ran generic_config_updater/test_dynamic_acl.py with and without the fix on multiple topos. Index out of bounds failure not occurring with the fix.

Any platform specific information?
I see that recently one LAG port has been added to c448o16. Internally we are using this patch to get our pass rate up but we may need to remove those topos from this change or create a follow-up PR when/if that change gets back ported to 202412.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>