Retry SSH connection if fails after reboot, with logging for easier debugging. (#17816)
What is the motivation for this PR?
To improve debugging by adding logs and retries for SSH connection failures after a reboot.

How did you do it?
Adds a retry with a quick timeout and no expected "search_regex".
Adds logs to capture ping status.
Raises an exception if the first connection attempt fails.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>