Snappi-Changed logging from info to debug for ECN packet marking (#18705)
What is the motivation for this PR?
The read_pcap.py file has function "is_ecn_marked" which is checking if the packet is ECN marked or not.

However the logging is set to INFO level, which means that every packet in the PCAP when checked for ECN marking, will generate an INFO level log. For check of 1000 packets, it will generate 1000 such INFO level logs.

How did you do it?
Reduced the severity of the log from INFO->DEBUG.

How did you verify/test it?
Ran ECN marking testcase. In 'Before' output, since 9k packets are being checked for ECN marked, it will generate 9k INFO level logs. In 'After' logs, the test-case itself is printing if the packet is detected as ECN marked or not.

Signed-off-by: Yael Tzur <ytzur@nvidia.com>