Add test for verifying ASCII output from console commnads (#18386)
What is the motivation for this PR?
Want to help verify the console stability of testbeds.
Sometimes if console is misconfigured, it will output random chars

How did you do it?
Follow the steps of other console tests.
Send common commands through console, and check if is ASCII output.

How did you verify/test it?
Verified by running the test on internal testbeds.

Signed-off-by: yenlu <yenlu@cisco.com>