[202411] Support Ubuntu 24 server in KVM (#18929)
What is the motivation for this PR?
When running with Ubuntu 24.04 host, many tasks from a sonic-mgmt container may fail.
One reason is that a newer version of Python being used in Ubuntu 24.04 that doesn't support some of the packages that ansible modules attempt to import. Such as imp module is removed in Python 3.11, need to use importlib, and
PR #16039 and #17883 did the fix and need to cherry-pick to 202411

How did you do it?
Merge the change of #16039 and #17883 to 202411