test_fib.py: test_ecmp_group_member_flap fails when selected source port is shut (#18808)
What is the motivation for this PR?
To fix failing test case test_fib.py: test_ecmp_group_member_flap

How did you do it?
To avoid selecting the port that is shutdown, in the fix it is being removed from the set of possible ingress ports.

How did you verify/test it?
Verified that the test case always passes with this fix.

Signed-off-by: yenlu <yenlu@cisco.com>