Refactor auditd testcases. (#18967)
Summary:

This PR addresses recent changes in the audit rule logic and the method for retrieving checksums from both containers. The goal is to align the implementation with the refactored architecture and ensure consistent behavior across environments.

Approach
What is the motivation for this PR?
The motivation behind this change is to support the recent refactor of the audit rule system and to unify the checksum retrieval process across containers. These updates aim to improve maintainability, reduce redundancy, and enhance the reliability of audit tracking.

How did you do it?
How did you verify/test it?
DUT verified.

Signed-off-by: yenlu <yenlu@cisco.com>