Remove special characters (special quotes) from Python scripts
Remove non-standard quote marks from two Python scripts in the ansible
directory. For Python 3, things will still work. However, in cases where
these scripts are executed in a Python 2 context (because the SONiC
device is running an old version of SONiC), it will cause problems. This
may be the case for upgrade path testing, where old versions of SONiC
might be running on the device.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>