chore: fix link oper down (#18975)
Description of PR
Summary: Due to the new API changes, we're not using the variable link_state anymore. During recovery, we checked for this variable to perform recovery, however this is never set.

Fixes # (issue)

Approach
What is the motivation for this PR?
Fix the link oper down on IXIA test case

Signed-off-by: Austin Pham <austinpham@microsoft.com>
Signed-off-by: yenlu <yenlu@cisco.com>