[CI]Refine pr test job timeout settings (#18989)
Description of PR
[CI]Refine pr test job timeout settings

Due to the increasing size of the test set and the limitation of test resources, the timeout of the original pr test pipeline is no longer sufficient. Many test jobs are almost completed, but are cancelled due to timeouts and then retried, resulting in unstable test results and waste of resources.

This PR extracts pr test pipeline timeout settings as a parameter for better control. It also adds variables to dynamically adjust the lock TB timeout and the overall test plan timeout.

Microsoft ADO 33280967:
Note:
In Azure Pipelines, the timeoutInMinutes setting does not support runtime variables like $(...) , it must be set using a literal integer value or a compile-time expression (i.e. using ${{ }} with parameters or variables known at compile time). So we cannot control it dynamically by settings in library groups. As well as template.

signed-off-by: jianquanye@microsoft.com