Merged PR 13314: Pass --ceos_neighbor_lacp_multiplier into 7060 nightly upgrade tests
This is for pipeline "7060cx.t0.202411" which can be found here:
https://dev.azure.com/mssonic/internal/_build?definitionId=2701

Note: I don't think a backport is necessary for this as we run the pipeline for 202411 out of the internal branch.

Nightly ADO: https://msazure.visualstudio.com/One/_workitems/edit/33173114

----
#### AI description  (iteration 1)
#### PR Classification
This PR updates the test configuration for the 7060 nightly upgrade tests.

#### PR Summary
The changes modify the pipeline YAML to pass the `--ceos_neighbor_lacp_multiplier=5` parameter during test execution.
- Updated `/.azure-pipelines/elastictest/arista/7060cx.t0.202411.yml` by appending `--ceos_neighbor_lacp_multiplier=5` to the metadata-scripts parameter.
- Added new parameter entries for `upgrade_path` and `platform_tests` to include the same flag.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->