Enable platform api watchdog  tests for dpu (#18189)
Description of PR
Enable platform api watchdog tests for dpu. Since platform api tests need to run a http server on the dut, so when dut is dpu, we need to add a nat config for the specific api service port for the tested dpu.

The relevant PRs:

[Mellanox][Nvidia-BF] Enable Watchdog support for Smartswitch DPU sonic-buildimage#22046
[Nvidia-Bluefield] [DPU] General DPU watchdog fixes sonic-buildimage#22587

Approach
What is the motivation for this PR?
Enable platform api watchdog tests for dpu.

How did you do it?
Add a nat config for the specific api service port for the tested dpu.

How did you verify/test it?
Run the platform api watchdog tests on dpu

Any platform specific information?
dpu

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>