adding fix for orc_timeout (#18833)
Description of PR
Previously, the timeout increase was applied only for modular chassis platforms. However, with the onboarding of several other platforms, we now need to implement this check in a platform-agnostic manner.

Summary:
Fixes # (issue)

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement
Back port request
 202205
 202305
 202311
 202405
 202411
 202505
Approach
What is the motivation for this PR?
Increasing orch_cpu_timeout to 120

How did you do it?
By adding a condition

Co-authored-by: yatishkoul <yatishkoul@gmail.com>
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>