consider the case where the portchannel is not an upstream port (#18908)
Approach
What is the motivation for this PR?
When running tests using the t2_single_node_max topology, and enable MACsec, we encountered failures in test_dataplane test. The root cause was using requester = upstream_links[portchannels[i]["members"][0]] to get upstream_links, while some portchannels link to LT2 and is not in upstream_links. Use this way can't get proper upstream.

How did you do it?
When iterating through portchannels, check if the current portchannel is in upstream_links. If it is not, skip the subsequent tests.

How did you verify/test it?
Tested the macsec/test_dataplane.py script locally using the t2_single_node_man topology with MACsec enabled and confirmed that the tests passed successfully.

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>