Merged PR 12946: Advertised routes count fix after AZNG rollback
After azng_migration -r, default route and loopback IP addresses are advertised to AH. The test checks for advertised route count to be same as number of DUTs which is incorrect.  It should be 1 more than the number of DUTs (for default route)

This currently works on chassis due to supervisor node being counted in len(duthosts), but no loopback advertised for sup.

Verified change on chassis and ut2