Roll back ceos image version since it's causing bgp failure (#19180)
What is the motivation for this PR?
Previously we updated ceos image version to 4.32.5M to meet the needs of test server OS version upgrade to Ubuntu 24, but it's causing bgp gr failure and the issue may not be able to have a quick fix.

How did you do it?
Roll back ceos image version to 4.29.3M.