Multi-device multi-tier testbed HLD (#18630)
What is the motivation for this PR?
Our current testbed topo design does not work well with multi-device multi-tier testbed setup, especially for the ones that does not require any virtualized neighbors and PTF.

How did you do it?
This document provides a comprehensive guide for building a SONiC testbed that supports using multiple devices in multi-tier topology as the test target (Network Under Test, NUT), without any neighbors such as cEOS. This testbed is particularly helpful to support test cases that doesn't require any virtual neighbors to be created to help with traffic routing and use hardware traffic generators to generate traffic.

Signed-off-by: yenlu <yenlu@cisco.com>