[BGP][topo] Modify leaf_number in ansible/vars/topo_t1-isolated-d510u2s2.yml (#19262)
What is the motivation for this PR?
Update leaf_number in topo)t1-isolated-d510u2s2.yml to 512. This number represents the number of t1 loopback routes that the t0 neighbor will advertise to the DUT.

How did you do it?
Update leaf_number in topo)t1-isolated-d510u2s2.yml to 512.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>