Ignore loganaylzer errors about the leapsecond file being expired
Depending on the image that is running, when it was built, and what
version of packages were used, the leapsecond fil might be expired.
Ignore errors from ntpd about it, since the only functional impact that
it might cause is that if there is a leapsecond, the system clock will
be off by a second for a little bit until it gets corrected.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>