change cpu cores allocation for kvm t2 (#19239)
Summary: Previously, we assigned an excessive amount of vCPU cores for every VM in kvm-t2 testbed and rely on QEMU to dynamically allocate CPU resources on host. However, we found that this may result in that QEMU allocates too much CPU time to the VMs and leave the host starved on CPU resources. Thus, we should restrict the number of vCPU cores allocated to each VM in kvm-t2.