Fix flakiness in pfcwd/test_pfcwd_cli.py (#17411) (#17619)
What is the motivation for this PR?
Sometimes the test ends up picking an egress interface which happens to be a member of a LAG. If the LAG has multiple members and only one of them is stormed the drop/forwards expectations don't take into account lag hashing.

Some of the traffic is hashed to another LAG member which is not being stormed and no drops will occur.

How did you do it?
The proposed fix is to shutdown the remaining LAG members.

How did you verify/test it?
Tested on Arista-7260CX3 with dualtor-120 topology and 202411 image.

Co-authored-by: Vivek Verma <137406113+vivekverma-arista@users.noreply.github.com>