[master][sonic-mgmt]Fix decap/test_subnet_decap.py::test_vlan_subnet_decap (#19061)
What is the motivation for this PR?
Fixes failures for decap/test_subnet_decap::test_vlan_subnet_decap[negative-IPv4/6] for active-standby dualtor topos
The test selects the DUT randomly, irrespective of it being the standby ToR
In the positive case, this will not be a problem - the encapsulated packet is sent with an outer SIP that matches the
config on the DUT, the packet gets decapped and gets routed towards the upstream T1s
In the negative case, the outer SIP does not match the config on the DUT and hence will not be decapsulated, but
forwarded to the outer DIP, which is set to IP 192.168.0.200/fc02:1000::200
This will require forwarding over one of the downstream ports where the ARP responder is also set up - if the
DUT is a standby; this forwarding will not happen
How did you do it?
This fix calls the fixture to ensure the selected DUT will be the active ToR, and the non-decapsulated packet can be
forwarded along the downstream port

How did you verify/test it?
Ran decap/test_subnet_decap.py on DCS-7050CX3

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>